Move the FIO reducer to the right place
